Move scan functions and data into their own file
